Move `EncryptedSolution` (and friends) to `publish` subdirectory
