+++ .vitepress/theme/index.ts
@@ -1,7 +1,7 @@
 import type { Theme } from 'vitepress';
 import MaterialFAQ from '../../src/components/MaterialFAQ.vue';
-import EncryptedSolution from '../../src/components/solution/EncryptedSolution.vue';
-import EncryptedSolutionGroup from '../../src/components/solution/EncryptedSolutionGroup.vue';
+import EncryptedSolution from '../../src/components/solution/publish/EncryptedSolution.vue';
+import EncryptedSolutionGroup from '../../src/components/solution/publish/EncryptedSolutionGroup.vue';
 import Layout from './Layout.vue';
 import './style/style.scss';
