+++ src/components/solution/publish/EncryptedSolutionGroup.vue
