+++ src/components/solution/publish/util/SolutionLock.vue
