+++ src/page/solution/[path].md
@@ -1,5 +1,5 @@
 <script setup>
-import SolutionRepositoryPage from '@/components/SolutionRepositoryPage.vue'
+import SolutionRepositoryPage from '@/components/solution/SolutionRepositoryPage.vue'
 </script>
 
 <SolutionRepositoryPage :requestedPage="$params.path" />
